Error encountered resolving symbol values statically in app-routing.module. SOLVED!
The solution to previously detected error is deleting all reference to
InventoryComponent in the router due to a stand-alone module no needs to
be declared in the router file, only needs to be imported and declared
in the app.module. All childs of InventoryModule are autoreferenced
inside the module itself.